+++ packages/openfort-react/src/components/ConnectModal/index.tsx
@@ -18,6 +18,7 @@ import BuyProcessing from '../Pages/BuyProcessing'
 import BuyProviderSelect from '../Pages/BuyProviderSelect'
 import BuySelectProvider from '../Pages/BuySelectProvider'
 import Connected from '../Pages/Connected'
+import ConnectedSuccess from '../Pages/ConnectedSuccess'
 import Connectors from '../Pages/Connectors'
 import CreateGuestUserPage from '../Pages/CreateGuestUserPage'
 import CreateWallet from '../Pages/CreateWallet'
@@ -25,9 +26,10 @@ import DownloadApp from '../Pages/DownloadApp'
 import EmailLogin from '../Pages/EmailLogin'
 import EmailOTP from '../Pages/EmailOTP'
 import EmailVerification from '../Pages/EmailVerification'
+import ExportKey from '../Pages/ExportKey'
 import ForgotPassword from '../Pages/ForgotPassword'
 import LinkEmail from '../Pages/LinkEmail'
-import LinkedProvidersPage from '../Pages/LinkedProviders'
+import LinkedProviders from '../Pages/LinkedProviders'
 import Loading from '../Pages/Loading'
 import LoadWallets from '../Pages/LoadWallets'
 import MobileConnectors from '../Pages/MobileConnectors'
@@ -64,7 +66,7 @@ const ConnectModal: React.FC<{
   const closeable = !(context.uiConfig.enforceSupportedChains && isConnected && !chainIsSupported)
 
   // const mainRoutes: ValueOf<typeof routes>[] = [
-  //   routes.PROFILE,
+  //   routes.CONNECTED,
   //   routes.LOADING,
   //   routes.PROVIDERS,
   //   routes.EMAIL_VERIFICATION,
@@ -74,7 +76,7 @@ const ConnectModal: React.FC<{
 
   // const showBackButton = (closeable && !mainRoutes.includes(route)) || (closeable && route === routes.PROVIDERS && user)
 
-  const _showInfoButton = closeable && route !== routes.PROFILE
+  const _showInfoButton = closeable && route !== routes.CONNECTED
 
   // const onBack = context.onBack
   //   ? context.onBack
@@ -90,12 +92,12 @@ const ConnectModal: React.FC<{
   //       }
 
   //       if (route === routes.CONNECTORS && user) {
-  //         context.setRoute(routes.PROFILE)
+  //         context.setRoute(routes.CONNECTED)
   //         return
   //       }
 
   //       if (route === routes.PROVIDERS || route === routes.SWITCHNETWORKS) {
-  //         context.setRoute(routes.PROFILE)
+  //         context.setRoute(routes.CONNECTED)
   //         return
   //       }
 
@@ -112,7 +114,7 @@ const ConnectModal: React.FC<{
     about: <About />,
     loading: <Loading />,
     loadWallets: <LoadWallets />,
-    connected: <Connected />,
+    connectedSuccess: <ConnectedSuccess />,
 
     createGuestUser: <CreateGuestUserPage />,
     socialProviders: <SocialProviders />,
@@ -134,9 +136,10 @@ const ConnectModal: React.FC<{
 
     providers: <Providers />,
     connect: <ConnectUsing />,
+    connected: <Connected />,
     profile: <Profile />,
-    linkedProviders: <LinkedProvidersPage />,
     switchNetworks: <SwitchNetworks />,
+    linkedProviders: <LinkedProviders />,
     connectWithMobile: <ConnectWithMobile />,
 
     noAssetsAvailable: <NoAssetsAvailable />,
@@ -152,6 +155,8 @@ const ConnectModal: React.FC<{
     buyProviderSelect: <BuyProviderSelect />,
     receive: <Receive />,
     buy: <Buy />,
+
+    exportKey: <ExportKey />,
   }
 
   function hide() {