+++ packages/openfort-react/src/hooks/openfort/useUI.ts
@@ -18,7 +18,7 @@ const safeRoutes: {
     routes.MOBILECONNECTORS,
   ],
   connected: [
-    routes.PROFILE,
+    routes.CONNECTED,
     { route: routes.CONNECTORS, connectType: 'linkIfUserConnectIfNoUser' },
     routes.SWITCHNETWORKS,
     routes.PROVIDERS,
@@ -104,19 +104,19 @@ export function useUI() {
     else if (!user) setRoute(routes.PROVIDERS)
     else if (!isConnected) setRoute(routes.LOAD_WALLETS)
     else if (needsRecovery) setRoute(routes.LOAD_WALLETS)
-    else setRoute(routes.PROFILE)
+    else setRoute(routes.CONNECTED)
   }
 
   const gotoAndOpen = (route: ValidRoutes) => {
     let validRoute: ValidRoutes = route
 
     if (!allRoutes.includes(route)) {
-      validRoute = isConnected ? routes.PROFILE : routes.PROVIDERS
+      validRoute = isConnected ? routes.CONNECTED : routes.PROVIDERS
       logger.log(`Route ${route} is not a valid route, navigating to ${validRoute} instead.`)
     } else {
       if (isConnected) {
         if (!safeRoutes.connected.includes(route)) {
-          validRoute = routes.PROFILE
+          validRoute = routes.CONNECTED
           logger.log(`Route ${route} is not a valid route when connected, navigating to ${validRoute} instead.`)
         }
       } else {
@@ -137,7 +137,7 @@ export function useUI() {
     close: () => setOpen(false),
     setIsOpen: setOpen,
 
-    openProfile: () => gotoAndOpen(routes.PROFILE),
+    openProfile: () => gotoAndOpen(routes.CONNECTED),
     openSwitchNetworks: () => gotoAndOpen(routes.SWITCHNETWORKS),
     openProviders: () => gotoAndOpen(routes.PROVIDERS),
     openWallets: () => gotoAndOpen({ route: routes.CONNECTORS, connectType: 'linkIfUserConnectIfNoUser' }),