+++ packages/openfort-react/src/index.ts
@@ -40,9 +40,9 @@ export { UserWallet, useWallets } from './hooks/openfort/useWallets'
 export { useChainIsSupported } from './hooks/useChainIsSupported'
 export { useChains } from './hooks/useChains'
 export { useOpenfortCore as useOpenfort } from './openfort/useOpenfort'
+export type { CustomTheme } from './styles/customTheme'
 export type {
   CustomAvatarProps,
-  CustomTheme,
   Languages,
   Mode,
   OpenfortHookOptions,