+++ packages/openfort-react/src/styles/themes/rounded.ts
@@ -1,4 +1,6 @@
-export default {
+import type { CustomTheme } from '../customTheme'
+
+const roundedTheme: CustomTheme = {
   '--ck-font-family':
     '"Nunito",ui-rounded,-apple-system,BlinkMacSystemFont,"Segoe UI",Helvetica,"Apple Color Emoji",Arial,sans-serif,"Segoe UI Emoji","Segoe UI Symbol"',
   '--ck-border-radius': '24px',
@@ -42,6 +44,9 @@ export default {
   '--ck-secondary-button-hover-background': '#F5F7F9',
   '--ck-secondary-button-hover-box-shadow':
     'inset 0 0 0 2px #DFE4EC, inset  0 -2px 0 0 #DFE4EC, 0px 2px 4px rgba(0, 0, 0, 0.02)',
+
+  '--ck-input-background': '#F5F7F9',
+  '--ck-input-hover-background': '#ffffff',
 
   '--ck-focus-color': '#1A88F8',
   '--ck-modal-box-shadow': `0px 3px 16px rgba(0, 0, 0, 0.08)`,
@@ -132,3 +137,5 @@ export default {
 
   '--ck-siwe-border': '#DFE4EC',
 }
+
+export default roundedTheme