+++ packages/openfort-react/src/types.ts
@@ -1,9 +1,10 @@
 import type { Languages as Lang } from './localizations'
+import type { CustomTheme } from './styles/customTheme'
+export type { CustomTheme }
 export type Languages = Lang
 
 export type Theme = 'auto' | 'web95' | 'retro' | 'soft' | 'midnight' | 'minimal' | 'rounded' | 'nouns'
 export type Mode = 'light' | 'dark' | 'auto'
-export type CustomTheme = any // OLD_TODO: define type
 
 export type All = {
   theme?: Theme